# docs/swagger.yaml
basePath: /
definitions:
  dto.ChatMessageRequest:
    properties:
      content:
        description: 消息内容
        type: string
      session_id:
        description: 聊天会话ID
        type: integer
    required:
    - content
    - session_id
    type: object
  dto.ChatStartRequest:
    properties:
      title:
        description: 聊天标题（可选）
        type: string
    type: object
...
  dto.SendMessageRequest:
    properties:
      message:
        type: string
      user_id:
        type: string
    required:
    - message
    - user_id
    type: object
...
  vo.ChatHistoryResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/vo.ChatMessageResponse'
        type: array
      session:
        $ref: '#/definitions/vo.ChatSessionResponse'
    type: object
  vo.ChatMessageResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      role:
        type: string
      session_id:
        type: integer
    type: object
  vo.ChatSessionResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  vo.ErrorResponse:
    properties:
      error:
        description: Error 错误信息
        example: invalid request
        type: string
    type: object
  vo.MessageListResponse:
    properties:
      limit:
        type: integer
      messages:
        items:
          $ref: '#/definitions/vo.MessageResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  vo.MessageResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      processed_at:
        type: string
      status:
        type: string
      task_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API支持
    url: http://www.swagger.io/support
  description: 这是一个用户排名管理系统的API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 排名系统API
  version: "1.0"
paths:
  /api/chat/{session_id}:
    get:
      description: 获取指定会话的聊天历史
      parameters:
      - description: 会话ID
        in: path
        name: session_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ChatHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取聊天历史
      tags:
      - chat
  /api/chat/send:
    post:
      consumes:
      - application/json
      description: 向指定会话发送消息，异步调用AI并通过WebSocket返回
      parameters:
      - description: 消息内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ChatMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 发送消息给AI
      tags:
      - chat
  /api/chat/sessions:
    get:
      description: 获取当前用户的所有聊天会话列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vo.ChatSessionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取用户所有聊天会话
      tags:
      - chat
  /api/chat/start:
    post:
      consumes:
      - application/json
      description: 创建一个新的AI聊天会话
      parameters:
      - description: 聊天标题
        in: body
        name: data
        schema:
          $ref: '#/definitions/dto.ChatStartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ChatSessionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/vo.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 开始新的聊天会话
      tags:
      - chat
...
  /api/messages:
    get:
      description: 获取指定用户的历史消息记录，支持分页和状态筛选
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 状态筛选
        enum:
        - pending
        - completed
        - failed
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.MessageListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取历史消息列表
      tags:
      - message
  /api/query-result:
    get:
      description: 根据任务ID查询延迟消息结果
      parameters:
      - description: 任务ID
        in: query
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 查询消息结果
      tags:
      - message
...
  /api/send-message:
    post:
      consumes:
      - application/json
      description: 发送一段消息，配置时间后返回
      parameters:
      - description: 消息内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 发送延迟消息
      tags:
      - message
schemes:
- http
swagger: "2.0"